- Better handling of errors - refactor util functions out into own files
