+++ resources/html/index.html
@@ -23,7 +23,7 @@
 </head>
 <body>
 <div class="editorContainer">
-    <textarea id="mainTextArea">// Enter your code here</textarea>
+    <textarea id="mainTextArea"></textarea>
 </div>
 <div class="controls">
     <select id="languageSelect">
