+++ src/code-editor.js
@@ -1,6 +1,7 @@
 "use strict";
 
 import * as config from "./config";
+import Utils from "./utils";
 
 
 /**
@@ -39,6 +40,11 @@ export default class CodeEditor {
         let lang = this.codeMirror.getOption("mode");
     }
 
+    setContent(content) {
+        this.options.value = content;
+        this.codeMirror.setOption("value", content);
+    }
+
     /**
      * Checks that the value has changed since the last save
      * and if so will make a post request to backend to save snippet
@@ -54,7 +60,7 @@ export default class CodeEditor {
         else {
             this.lastValue = snippet;
             return new Promise(function(resolve, reject) {
-                CodeEditor.xmlReq(
+                Utils.xmlReq(
                     "/save-snippet",
                     "POST",
                     {language: language, snippet: snippet})
@@ -75,39 +81,4 @@ export default class CodeEditor {
 
     }
 
-    /**
-     * Static method that wraps XMLHttpRequest in a promise
-     * @param {string} url - url to make the request to
-     * @param {'POST'|'GET'} type
-     * @param {object} params - optional params to pass
-     * @returns {Promise}
-     */
-    static xmlReq(url, type, params={}) {
-        return new Promise(function(resolve, reject) {
-            let xhttp = new XMLHttpRequest();
-            xhttp.onload = function() {
-                if (xhttp.readyState === 4 && xhttp.status === 200) {
-                    let response = JSON.parse(xhttp.responseText);
-                    if (response.status == config.RESPONSE_SUCCESS_STATUS) {
-                        resolve(response.response);
-                    } else {
-                        // Reject entire response. Calling method can deal with error
-                        reject(response);
-                    }
-                }
-            };
-            xhttp.onerror = function(err) {
-                reject(err);
-            };
-            xhttp.open(type, url, true);
-            xhttp.setRequestHeader("Content-type", "application/x-www-form-urlencoded");
-
-            let queryStr = Object.keys(params).reduce(function(a, k) {
-                a.push(k + "=" + encodeURIComponent(params[k]));
-                return a;
-            }, []).join("&");
-
-            xhttp.send(queryStr);
-        });
-    }
 }
