+++ src/config.js
@@ -13,6 +13,19 @@ export const SNIPPET_QUERY_PARAM = "s";
 export const DEFAULT_LANG =  "javascript";
 
 /**
+ * The default content of the editor if we are creating
+ * a new editor
+ * @type {string}
+ */
+export const DEFAULT_CONTENT = "// Enter your code here :)";
+
+/**
+ * String to display when we cant find users requested snippet
+ * @type {string}
+ */
+export const SNIPPET_NOT_FOUND = "// Sorry We can't find the snipped you are looking for\n// Why not create a new one :)";
+
+/**
  * The status the server returns when request is succdessful
  * @type {string}
  */
@@ -23,3 +36,4 @@ export const RESPONSE_SUCCESS_STATUS = "ok";
  * @type {string}
  */
 export const RESPONSE_ERROR_STATUS = "error";
+
